refactor(client): isolate persisted state and runtime utilities
