+++ packages/client/src/authorization-types.ts
@@ -1,0 +1,7 @@
+export interface MdbaseDeviceAuthorization {
+  userCode: string;
+  verificationUri: string;
+  verificationUriComplete: string;
+  expiresAt: number;
+  intervalSeconds: number;
+}
